requirements.txt
blinker==1.9.0
click==8.1.7
colorama==0.4.6
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python==4.10.0.84
packaging==24.2
pillow==11.0.0
pillow_heif==0.20.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
